requirements.txt
Scrapy==2.8.0
python-telegram-bot==20.1
SQLAlchemy==2.0.5.post1
SQLAlchemy-Utils==0.40.0
psycopg2==2.9.5